requirements.txt
httpx
click
pycryptodome
pycryptodomex